design view desktop and movil

+++ core/app/layouts/layout.php
@@ -237,7 +237,9 @@ if( $q==0 ||  $q<=$product->inventary_min){
                 <li><a href="./?view=cotizations">Cotizaciones</a></li> 
                 <li><a href="./?view=orders-approved">Pedidos</a></li>
                 <li><a href="./?view=cotizationscancel">Cotizaciones Canceladas</a></li>
-                <li><a href="./?view=cotizationsdelete">Cotizaciones Eliminadas</a></li>
+                <?php if (isset($_SESSION['is_admin'])): ?>
+                  <li><a href="./?view=cotizationsdelete">Cotizaciones Eliminadas</a></li>
+                <?php endif ?>
 
               </ul>
             </li>
